refactor: move widgets to objects & projects dirs
