+++ lab/models.py
@@ -1,3 +1,5 @@
+# pylint: disable=unused-import
+
 from .objects import Location, Object, ObjectGroup, Period  # noqa: F401
 from .participations.models import Institution, Participation  # noqa: F401
 from .projects.models import BeamTimeRequest, Project  # noqa: F401
